requirements.txt
lxml==5.3.0
jupyterlab==4.2.5
notebook==6.1.5
absl_py==2.1.0
accelerate==0.30.1
addict==2.4.0
easydict==1.13
einops==0.8.0
filelock==3.16.1
flash_attn==2.5.9.post1
Flask==3.0.3
groq==0.11.0
gym==0.26.2
h5py==3.12.1
hydra-core==1.3.2
jsonlines==4.0.0
lmdb==1.4.1
matplotlib==3.9.2
msgpack_numpy==0.4.8
msgpack_python==0.5.6
numpy==1.26.4
omegaconf==2.3.0
open3d==0.18.0
open_clip_torch==2.24.0
opencv_python==4.10.0.82
Pillow==10.4.0
pytest==8.3.3
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
spconv_cu120==2.3.6
tensorboardX==2.6.2.2
termcolor==2.5.0
timm==1.0.9
torch_scatter==2.1.2
tqdm==4.66.4
transformers==4.41.2
typed_argument_parser==1.10.0
yacs==0.1.8